JSON DONE! (i think)
Fuck me, the mod now has 7212 json files...
I believe this is the end of the jsons for 1.2.
Fences and Fence Gates for non-stonecuttable materials, as well as Wood Plank Walls, arent obtainable in Survival for the time being. If I can figure out a decent way to craft them, it'll happen. I have an idea, but it's a bit ambitious, and not for this update.