Preliminary 1.3.0 work
- Added Corners (sideways stairs)
- Added Posts (fence posts that don't connect and can face all 3 axis)
TODO: Fix top and bottom textures being the side textures for relevant corners.
TODO: Add separate model for horizontal posts so as to have the correct textures on the correct sides.
TODO: Fix inventory rendering of posts based on vanilla fences
TODO: Add Columns. (like posts are to fences, columns are to walls)
TODO: All the other guffins that comes with adding things. Localisation, Recipes, Loot Tables and that.